Give transfers.min_transfer_time room for the longest fixed link
The time is in seconds, and a fixed link may take up to 999 minutes - 59940,
five digits. Declared as four, it is a smallint on Postgres, which stops at
32767, so a link longer than about nine hours rolled the transfers load back.
The column this replaced was an unsigned smallint on MySQL, which reaches
65535. Five digits is an integer on Postgres and a mediumint on MySQL.

+++ apps/dtd2mysql/src/gtfs/schema.ts
@@ -177,7 +177,8 @@ export const transfers = gtfsTable({
   from_trip_id: defaultTo(ascii(varchar(32)), ""),
   to_trip_id: defaultTo(ascii(varchar(32)), ""),
   transfer_type: integer(2),
-  min_transfer_time: nullable(integer(4)),
+  // seconds, and a fixed link may take up to 999 minutes: 59940 is five digits
+  min_transfer_time: nullable(integer(5)),
   mode: nullable(varchar(255)),
   start_time: nullable(time),
   end_time: nullable(time),
